clean

+++ lib/ESP8266System/Fsm.h
@@ -17,7 +17,7 @@ class Fsm {
     void addTransition(State* stateFrom, State* stateTo, u8 event, std::function<void(void)> onTransition);
     void addTimedTransition(State* stateFrom, State* stateTo, u64 interval, std::function<void(void)> onTransition);
     
-    void trigger(int event);
+    void trigger(u8 event);
     void loop();
   
   private:
